Fixed theming issue
