--- apps/web/app/layout.tsx
@@ -1,17 +1,0 @@
-"use client";
-import ResponsiveDrawer from "../../../packages/ui/NavBar";
-
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <html>
-      <body>
-        <ResponsiveDrawer window={undefined} />
-        {children}
-      </body>
-    </html>
-  );
-}
