--- packages/ui/themes/dark.theme.js
@@ -1,185 +1,0 @@
-import { createTheme } from "@mui/material";
-
-export const darkTheme = createTheme({
-  palette: {
-    mode: "dark",
-    // changes begin here, with custom names, feel free to change them if they don't work
-    background: {
-      default: "#141625",
-    },
-    //statics color for both themes
-    purple: {
-      main: "#7C5DFA",
-      hover: "#9277FF",
-    },
-    danger: {
-      softRed: "#EC5757",
-      veryLightRed: "#FF9797",
-      contrastText: "white",
-    },
-    status: {
-      paid: {
-        main: "#33D69F",
-        secondary: "rgba(51, 214, 159, 0.0571)",
-      },
-      pending: {
-        main: "#FF8F00",
-        secondary: "rgba(255, 143, 0, 0.0571)     ",
-      },
-
-      draft: {
-        main: "#DFE3FA",
-        secondary: "rgba(223, 227, 250, 0.0571)",
-      },
-    },
-    // mutable colors for themes
-    saveDraft: {
-      main: "#373B53",
-      hover: "#1E2139",
-      contrastText: "#DFE3FA",
-    },
-    edit: {
-      main: "#252945",
-      hover: "#FFFFFF",
-      contrastText: "#DFE3FA",
-    },
-    componentBackground: {
-      main: "#252945",
-      secondary: "#1E2139",
-    },
-    icon: {
-      date: "#7E88C3",
-    },
-    font: {
-      main: "#DFE3FA",
-    },
-    // changes finish here
-    // added this, we need to have a value for both themes, otherwise it doesn't exist
-    // and we have the "Cannot read properties of undefined (reading 'dark')" error
-    softBlue: {
-      main: "#7C5DFA",
-      contrastText: "white",
-    },
-    secondary: {
-      main: "#9277FF",
-    },
-    softRed: {
-      main: "#EC5757",
-    },
-    veryLightRed: {
-      main: "#FF9797",
-    },
-    lightGrayishBlue: {
-      main: "#252945",
-    },
-    lightGreyishBlueHover: {
-      main: "#DFE3FA",
-    },
-    veryDarkGrayishBlue: {
-      main: "#373B53",
-    },
-    white: {
-      main: "white",
-    },
-    mostlyBlack: {
-      main: "#0C0E16",
-    },
-    desaturatedBlue: {
-      main: "#1E2139",
-    },
-    slightlyDesaturatedBlue: {
-      main: "#7E88C3",
-    },
-    darkGreyishBlue: {
-      main: "#888EB0",
-    },
-    pureOrange: {
-      main: "#FF8F00",
-    },
-    limeGreen: {
-      main: "#33D69F",
-    },
-    navBg: {
-      main: "#1E2139",
-    },
-  },
-  typography: {
-    h1: {
-      fontSize: "32px",
-      fontWeight: "700",
-      lineHeight: "36px",
-      letterSpacing: "-1px",
-    },
-    h2: {
-      fontSize: "20px",
-      fontWeight: "700",
-      lineHeight: "22px",
-      letterSpacing: "-0.63px",
-    },
-    h3: {
-      fontSize: "16px",
-      fontWeight: "700",
-      lineHeight: "24px",
-      letterSpacing: "-0.8px",
-    },
-    h4: {
-      fontSize: "12px",
-      fontWeight: "700",
-      lineHeight: "15px",
-      letterSpacing: "-0.25px",
-    },
-    body1: {
-      fontSize: "12px",
-      fontWeight: "500",
-      lineHeight: "15px",
-      letterSpacing: "-0.25px",
-    },
-    body2: {
-      fontSize: "11px",
-      fontWeight: "500",
-      lineHeight: "18px",
-      letterSpacing: "-0.23px",
-    },
-    subtitle1: {
-      fontSize: "11px",
-      fontWeight: "500",
-      lineHeight: "18px",
-      letterSpacing: "-0.23px",
-      color: "#7E88C3",
-    },
-    overline: {
-      fontSize: "120px",
-      fontWeight: "700",
-      lineHeight: "137px",
-      letterSpacing: "-3.75px",
-    },
-    h5: {
-      fontSize: "20px",
-      fontWeight: "700",
-      lineHeight: "39px",
-      letterSpacing: "8px",
-      color: "#888EB0",
-    },
-    button: {
-      fontSize: "12px",
-      fontWeight: "700",
-      lineHeight: "15px",
-      letterSpacing: "-0.25px",
-      textTransform: "none",
-    },
-    poster: {
-      color: "white",
-    },
-    fontFamily: "League Spartan",
-    fontWeightMedium: "500",
-    fontWeightBold: "700",
-    textTransform: "none",
-  },
-  caption: {
-    fontSize: "12px",
-    fontWeight: "700",
-    lineHeight: "15px",
-    letterSpacing: "-0.25px",
-    textTransform: "none",
-  },
-});
